refactoring: reorganize analysis code etc. almost done
