+++ runner/frameworks/classification/analyze.py
@@ -1,53 +1,51 @@
 import json
-import logging
-import shutil
-from dataclasses import asdict
+from dataclasses import asdict, dataclass
 from pathlib import Path
-from typing import Any, Dict, List
+from typing import Optional
 
-import matplotlib.pyplot as plt
-import numpy as np
 import torch
-import wandb
-from matplotlib.figure import Figure
 from omegaconf import DictConfig
 
+from retinal_rl.analysis import channel_analysis as channel_ana
+from retinal_rl.analysis import receptive_fields
 from retinal_rl.analysis.plot import (
-    layer_receptive_field_plots,
+    FigureLogger,
     plot_brain_and_optimizers,
-    plot_channel_statistics,
-    plot_histories,
     plot_receptive_field_sizes,
-    plot_reconstructions,
-    plot_transforms,
 )
-from retinal_rl.analysis.statistics import (
-    CNNStatistics,
-    LayerStatistics,
-    cnn_statistics,
-    reconstruct_images,
+from retinal_rl.analysis.reconstructions import perform_reconstruction_analysis
+from retinal_rl.analysis.transforms_analysis import (
+    plot_transforms,
     transform_base_images,
 )
 from retinal_rl.classification.imageset import Imageset
 from retinal_rl.models.brain import Brain
-from retinal_rl.models.loss import ReconstructionLoss
 from retinal_rl.models.objective import ContextT, Objective
+from retinal_rl.util import FloatArray, NumpyEncoder
 
 ### Infrastructure ###
-
-
-logger = logging.getLogger(__name__)
 
 init_dir = "initialization_analysis"
 
 
-class NumpyEncoder(json.JSONEncoder):
-    """JSON encoder that handles numpy arrays."""
+@dataclass
+class AnalysesCfg:
+    run_dir: Path
+    plot_dir: Path
+    checkpoint_plot_dir: Path
+    data_dir: Path
+    use_wandb: bool
+    channel_analysis: bool
+    plot_sample_size: int
 
-    def default(self, obj: Any) -> Any:
-        if isinstance(obj, np.ndarray):
-            return obj.tolist()
-        return super().default(obj)
+    def __post_init__(self):
+        self.analyses_dir = Path(self.data_dir) / "analyses"
+
+        # Ensure all dirs exist
+        self.run_dir.mkdir(exist_ok=True)
+        self.plot_dir.mkdir(exist_ok=True)
+        self.checkpoint_plot_dir.mkdir(exist_ok=True)
+        self.analyses_dir.mkdir(exist_ok=True)
 
 
 ### Analysis ###
@@ -58,7 +56,7 @@ def analyze(
     device: torch.device,
     brain: Brain,
     objective: Objective[ContextT],
-    histories: Dict[str, List[float]],
+    histories: dict[str, list[float]],
     train_set: Imageset,
     test_set: Imageset,
     epoch: int,
@@ -66,71 +64,70 @@ def analyze(
 ):
     ## DictConfig
 
-    # Path creation
-    run_dir = Path(cfg.path.run_dir)
-    run_dir.mkdir(exist_ok=True)
-
-    plot_dir = Path(cfg.path.plot_dir)
-    plot_dir.mkdir(exist_ok=True)
-
-    checkpoint_plot_dir = Path(cfg.path.checkpoint_plot_dir)
-    checkpoint_plot_dir.mkdir(exist_ok=True)
-
-    analyses_dir = Path(cfg.path.data_dir) / "analyses"
-    analyses_dir.mkdir(exist_ok=True)
-
-    # Variables
-    use_wandb = cfg.logging.use_wandb
-    channel_analysis = cfg.logging.channel_analysis
-    plot_sample_size = cfg.logging.plot_sample_size
-
-    ## Analysis
-
-    if not use_wandb:
-        _plot_and_save_histories(plot_dir, histories)
-
-    # Get CNN statistics and save them
-    cnn_stats = cnn_statistics(
-        device,
-        test_set,
-        brain,
-        channel_analysis,
-        plot_sample_size,
+    _cfg = AnalysesCfg(
+        Path(cfg.path.run_dir),
+        Path(cfg.path.plot_dir),
+        Path(cfg.path.checkpoint_plot_dir),
+        Path(cfg.path.data_dir),
+        cfg.logging.use_wandb,
+        cfg.logging.channel_analysis,
+        cfg.logging.plot_sample_size,
+    )
+    log = FigureLogger(
+        _cfg.use_wandb, _cfg.plot_dir, _cfg.checkpoint_plot_dir, _cfg.run_dir
     )
 
-    # Save CNN statistics
-    with open(analyses_dir / f"cnn_stats_epoch_{epoch}.json", "w") as f:
-        json.dump(asdict(cnn_stats), f, cls=NumpyEncoder)
+    ## Analysis
+    log.plot_and_save_histories(histories)
 
-    if epoch == 0:
-        _perform_initialization_analysis(
-            channel_analysis,
-            use_wandb,
-            analyses_dir,
-            plot_dir,
-            checkpoint_plot_dir,
-            run_dir,
-            brain,
-            objective,
-            train_set,
-            cnn_stats,
-        )
+    # # Save CNN statistics # TODO: how to do this now...
+    # with open(_cfg.analyses_dir / f"cnn_stats_epoch_{epoch}.json", "w") as f:
+    #     json.dump(asdict(cnn_stats), f, cls=NumpyEncoder)
 
-    _analyze_layers(
-        channel_analysis,
-        use_wandb,
-        plot_dir,
-        checkpoint_plot_dir,
-        cnn_stats,
+    # perform different analyses
+    input_shape, rf_result = receptive_fields.analyze(brain, device)
+    receptive_fields.plot(
+        log,
+        rf_result,
         epoch,
         copy_checkpoint,
     )
 
-    _perform_reconstruction_analysis(
-        use_wandb,
-        analyses_dir,
-        plot_dir,
-        checkpoint_plot_dir,
+    if _cfg.channel_analysis:
+        spectral_result = channel_ana.spectral_analysis(
+            device, test_set, brain, _cfg.plot_sample_size
+        )
+        histogram_result = channel_ana.histogram_analysis(
+            device, test_set, brain, _cfg.plot_sample_size
+        )
+        # TODO: Do we really want to replot rfs here?
+        channel_ana.plot(
+            log,
+            rf_result,
+            spectral_result,
+            histogram_result,
+            epoch,
+            copy_checkpoint,
+        )
+    else:
+        spectral_result, histogram_result = None, None
+
+    # plot results
+    if epoch == 0:
+        _default_initialization_plots(log, brain, objective, input_shape, rf_result)
+        _extended_initialization_plots(
+            log,
+            _cfg.channel_analysis,
+            _cfg.analyses_dir,
+            train_set,
+            rf_result,
+            spectral_result,
+            histogram_result,
+        )
+
+    perform_reconstruction_analysis(
+        log,
+        _cfg.analyses_dir,
         device,
         brain,
         objective,
@@ -140,61 +137,48 @@ def analyze(
         copy_checkpoint,
     )
 
-    hist_fig = plot_histories(histories)
-    _save_figure(plot_dir, "", "histories", hist_fig)
-    plt.close(hist_fig)
+    log.plot_and_save_histories(histories, save_always=True)
 
 
-def _plot_and_save_histories(plot_dir: Path, histories: Dict[str, List[float]]):
-    hist_fig = plot_histories(histories)
-    _save_figure(plot_dir, "", "histories", hist_fig)
-    plt.close(hist_fig)
-
-
-def _perform_initialization_analysis(
-    channel_analysis: bool,
-    use_wandb: bool,
-    analyses_dir: Path,
-    plot_dir: Path,
-    checkpoint_plot_dir: Path,
-    run_dir: Path,
+def _default_initialization_plots(
+    log: FigureLogger,
     brain: Brain,
     objective: Objective[ContextT],
-    train_set: Imageset,
-    cnn_stats: CNNStatistics,
+    input_shape: tuple[int, ...],
+    rf_result: dict[str, FloatArray],
 ):
-    summary = brain.scan()
-    filepath = run_dir / "brain_summary.txt"
-    filepath.write_text(summary)
+    log.save_summary(brain)
 
-    if use_wandb:
-        wandb.save(str(filepath), base_path=run_dir, policy="now")
-
+    # TODO: Move this somewhere accessible for RL
     # TODO: This is a bit of a hack, we should refactor this to get the relevant information out of  cnn_stats
-    rf_sizes_fig = plot_receptive_field_sizes(**asdict(cnn_stats))
-    _process_figure(
-        use_wandb,
-        plot_dir,
-        checkpoint_plot_dir,
-        False,
+    rf_sizes_fig = plot_receptive_field_sizes(input_shape, rf_result)
+    log.log_figure(
         rf_sizes_fig,
         init_dir,
         "receptive_field_sizes",
         0,
+        False,
     )
 
     graph_fig = plot_brain_and_optimizers(brain, objective)
-    _process_figure(
-        use_wandb,
-        plot_dir,
-        checkpoint_plot_dir,
-        False,
+    log.log_figure(
         graph_fig,
         init_dir,
         "brain_graph",
         0,
+        False,
     )
 
+
+def _extended_initialization_plots(
+    log: FigureLogger,
+    channel_analysis: bool,
+    analyses_dir: Path,
+    train_set: Imageset,
+    rf_result: dict[str, FloatArray],
+    spectral_result: Optional[dict[str, channel_ana.SpectralAnalysis]] = None,
+    histogram_result: Optional[dict[str, channel_ana.HistogramAnalysis]] = None,
+):
     transforms = transform_base_images(train_set, num_steps=5, num_images=2)
     # Save transform statistics
     transform_path = analyses_dir / "transforms.json"
@@ -202,232 +186,52 @@ def _perform_initialization_analysis(
         json.dump(asdict(transforms), f, cls=NumpyEncoder)
 
     transforms_fig = plot_transforms(**asdict(transforms))
-    _process_figure(
-        use_wandb,
-        plot_dir,
-        checkpoint_plot_dir,
-        False,
+    log.log_figure(
         transforms_fig,
         init_dir,
         "transforms",
         0,
+        False,
     )
 
-    _analyze_input_layer(
-        use_wandb,
-        plot_dir,
-        checkpoint_plot_dir,
-        cnn_stats.layers["input"],
-        channel_analysis,
-    )
-
-
-def _analyze_layers(
-    channel_analysis: bool,
-    use_wandb: bool,
-    plot_dir: Path,
-    checkpoint_plot_dir: Path,
-    cnn_stats: CNNStatistics,
-    epoch: int,
-    copy_checkpoint: bool,
-):
-    for layer_name, layer_data in cnn_stats.layers.items():
-        if layer_name != "input":
-            _analyze_regular_layer(
-                use_wandb,
-                plot_dir,
-                checkpoint_plot_dir,
-                layer_name,
-                layer_data,
-                epoch,
-                copy_checkpoint,
-                channel_analysis,
-            )
+    if spectral_result and histogram_result:
+        _analyze_input_layer(
+            log,
+            rf_result["input"],
+            spectral_result["input"],
+            histogram_result["input"],
+            channel_analysis,
+        )
 
 
 def _analyze_input_layer(
-    use_wandb: bool,
-    plot_dir: Path,
-    checkpoint_plot_dir: Path,
-    layer_statistics: LayerStatistics,
+    log: FigureLogger,
+    rf_result: FloatArray,
+    spectral_result: channel_ana.SpectralAnalysis,
+    histogram_result: channel_ana.HistogramAnalysis,
     channel_analysis: bool,
 ):
-    layer_rfs = layer_receptive_field_plots(layer_statistics.receptive_fields)
-    _process_figure(
-        use_wandb,
-        plot_dir,
-        checkpoint_plot_dir,
-        False,
+    layer_rfs = receptive_fields.layer_receptive_field_plots(rf_result)
+    log.log_figure(
         layer_rfs,
         init_dir,
         "input_rfs",
         0,
-    )
-
+        False,
+    )  # TODO: What's the purpose of it - it's just RGB I guess?
     if channel_analysis:
-        layer_dict = asdict(layer_statistics)
-        num_channels = int(layer_dict.pop("num_channels"))
-        for channel in range(num_channels):
-            channel_fig = plot_channel_statistics(
-                **layer_dict, layer_name="input", channel=channel
+        for channel in range(rf_result.shape[0]):
+            channel_fig = channel_ana.layer_channel_plots(
+                rf_result,
+                spectral_result,
+                histogram_result,
+                layer_name="input",
+                channel=channel,
             )
-            _process_figure(
-                use_wandb,
-                plot_dir,
-                checkpoint_plot_dir,
-                False,
+            log.log_figure(
                 channel_fig,
                 init_dir,
                 f"input_channel_{channel}",
                 0,
+                False,
             )
-
-
-def _analyze_regular_layer(
-    use_wandb: bool,
-    plot_dir: Path,
-    checkpoint_plot_dir: Path,
-    layer_name: str,
-    layer_statistics: LayerStatistics,
-    epoch: int,
-    copy_checkpoint: bool,
-    channel_analysis: bool,
-):
-    layer_rfs = layer_receptive_field_plots(layer_statistics.receptive_fields)
-    _process_figure(
-        use_wandb,
-        plot_dir,
-        checkpoint_plot_dir,
-        copy_checkpoint,
-        layer_rfs,
-        "receptive_fields",
-        f"{layer_name}",
-        epoch,
-    )
-
-    if channel_analysis:
-        layer_dict = asdict(layer_statistics)
-        num_channels = int(layer_dict.pop("num_channels"))
-        for channel in range(num_channels):
-            channel_fig = plot_channel_statistics(
-                **layer_dict, layer_name=layer_name, channel=channel
-            )
-
-            _process_figure(
-                use_wandb,
-                plot_dir,
-                checkpoint_plot_dir,
-                copy_checkpoint,
-                channel_fig,
-                f"{layer_name}_layer_channel_analysis",
-                f"channel_{channel}",
-                epoch,
-            )
-
-
-def _perform_reconstruction_analysis(
-    use_wandb: bool,
-    analyses_dir: Path,
-    plot_dir: Path,
-    checkpoint_plot_dir: Path,
-    device: torch.device,
-    brain: Brain,
-    objective: Objective[ContextT],
-    train_set: Imageset,
-    test_set: Imageset,
-    epoch: int,
-    copy_checkpoint: bool,
-):
-    reconstruction_decoders = [
-        loss.target_decoder
-        for loss in objective.losses
-        if isinstance(loss, ReconstructionLoss)
-    ]
-
-    for decoder in reconstruction_decoders:
-        norm_means, norm_stds = train_set.normalization_stats
-        rec_dict = asdict(
-            reconstruct_images(device, brain, decoder, train_set, test_set, 5)
-        )
-        # Save the reconstructions
-        rec_path = analyses_dir / f"{decoder}_reconstructions_epoch_{epoch}.json"
-        with open(rec_path, "w") as f:
-            json.dump(rec_dict, f, cls=NumpyEncoder)
-
-        recon_fig = plot_reconstructions(
-            norm_means,
-            norm_stds,
-            *rec_dict["train"].values(),
-            *rec_dict["test"].values(),
-            num_samples=5,
-        )
-        _process_figure(
-            use_wandb,
-            plot_dir,
-            checkpoint_plot_dir,
-            copy_checkpoint,
-            recon_fig,
-            "reconstruction",
-            f"{decoder}_reconstructions",
-            epoch,
-        )
-
-
-### Helper Functions ###
-
-
-def _save_figure(plot_dir: Path, sub_dir: str, file_name: str, fig: Figure) -> None:
-    dir = plot_dir / sub_dir
-    dir.mkdir(exist_ok=True)
-    file_path = dir / f"{file_name}.png"
-    fig.savefig(file_path)
-
-
-def _checkpoint_copy(
-    plot_dir: Path, checkpoint_plot_dir: Path, sub_dir: str, file_name: str, epoch: int
-) -> None:
-    src_path = plot_dir / sub_dir / f"{file_name}.png"
-
-    dest_dir = checkpoint_plot_dir / f"epoch_{epoch}" / sub_dir
-    dest_dir.mkdir(parents=True, exist_ok=True)
-    dest_path = dest_dir / f"{file_name}.png"
-
-    shutil.copy2(src_path, dest_path)
-
-
-def _wandb_title(title: str) -> str:
-    # Split the title by slashes
-    parts = title.split("/")
-
-    def capitalize_part(part: str) -> str:
-        # Split the part by dashes
-        words = part.split("_")
-        # Capitalize each word
-        capitalized_words = [word.capitalize() for word in words]
-        # Join the words with spaces
-        return " ".join(capitalized_words)
-
-    # Capitalize each part, then join with slashes
-    capitalized_parts = [capitalize_part(part) for part in parts]
-    return "/".join(capitalized_parts)
-
-
-def _process_figure(
-    use_wandb: bool,
-    plot_dir: Path,
-    checkpoint_plot_dir: Path,
-    copy_checkpoint: bool,
-    fig: Figure,
-    sub_dir: str,
-    file_name: str,
-    epoch: int,
-) -> None:
-    if use_wandb:
-        title = f"{_wandb_title(sub_dir)}/{_wandb_title(file_name)}"
-        img = wandb.Image(fig)
-        wandb.log({title: img}, commit=False)
-    else:
-        _save_figure(plot_dir, sub_dir, file_name, fig)
-        if copy_checkpoint:
-            _checkpoint_copy(plot_dir, checkpoint_plot_dir, sub_dir, file_name, epoch)
-    plt.close(fig)
